rename components to pureComponents, apply changes for PR review
