+++ src/pureComponents/Card.tsx
