+++ src/pureComponents/Modal.tsx
