+++ src/pureComponents/ProgrammingLanguageIcon.tsx
