+++ src/pureComponents/Sidebar/Box/CompanyItem.tsx
