+++ src/pureComponents/Sidebar/Box/FounderItem.tsx
