+++ src/pureComponents/Sidebar/Box/GithubStatItem.tsx
