+++ src/pureComponents/Sidebar/Box/index.tsx
