+++ src/pureComponents/Sidebar/Section/index.tsx
