+++ src/pureComponents/Sidebar/Small.tsx
