+++ src/pureComponents/Sidebar/index.tsx
